refactor: ui에서 사용되는 데이터를 ui.data로 이동
