+++ android/app/src/main/java/app/priceguard/data/repository/ProductRepositoryImpl.kt
@@ -2,20 +2,20 @@ package app.priceguard.data.repository
 
 import app.priceguard.data.GraphDataConverter
 import app.priceguard.data.dto.add.ProductAddRequest
-import app.priceguard.data.dto.add.ProductAddResult
-import app.priceguard.data.dto.detail.ProductDetailResult
-import app.priceguard.data.dto.list.ProductData
 import app.priceguard.data.dto.patch.PricePatchRequest
-import app.priceguard.data.dto.patch.PricePatchResult
-import app.priceguard.data.dto.recommend.RecommendProductData
-import app.priceguard.data.dto.renew.RenewResult
 import app.priceguard.data.dto.verify.ProductVerifyRequest
-import app.priceguard.data.dto.verify.ProductVerifyResult
 import app.priceguard.data.network.APIResult
 import app.priceguard.data.network.ProductAPI
 import app.priceguard.data.network.ProductErrorState
 import app.priceguard.data.network.ProductRepositoryResult
 import app.priceguard.data.network.getApiResult
+import app.priceguard.ui.data.PricePatchResult
+import app.priceguard.ui.data.ProductAddResult
+import app.priceguard.ui.data.ProductData
+import app.priceguard.ui.data.ProductDetailResult
+import app.priceguard.ui.data.ProductVerifyResult
+import app.priceguard.ui.data.RecommendProductData
+import app.priceguard.ui.data.RenewResult
 import javax.inject.Inject
 
 class ProductRepositoryImpl @Inject constructor(
