+++ android/app/src/main/java/app/priceguard/data/repository/UserRepositoryImpl.kt
@@ -1,14 +1,14 @@
 package app.priceguard.data.repository
 
 import app.priceguard.data.dto.login.LoginRequest
-import app.priceguard.data.dto.login.LoginResult
 import app.priceguard.data.dto.login.LoginState
 import app.priceguard.data.dto.signup.SignupRequest
-import app.priceguard.data.dto.signup.SignupResult
 import app.priceguard.data.dto.signup.SignupState
 import app.priceguard.data.network.APIResult
 import app.priceguard.data.network.UserAPI
 import app.priceguard.data.network.getApiResult
+import app.priceguard.ui.data.LoginResult
+import app.priceguard.ui.data.SignupResult
 import javax.inject.Inject
 
 class UserRepositoryImpl @Inject constructor(private val userAPI: UserAPI) : UserRepository {
