+++ android/app/src/main/java/app/priceguard/ui/data/LoginResult.kt
@@ -1,4 +1,6 @@
-package app.priceguard.data.dto.login
+package app.priceguard.ui.data
+
+import app.priceguard.data.dto.login.LoginState
 
 data class LoginResult(
     val loginState: LoginState,
