+++ android/app/src/main/java/app/priceguard/ui/data/ProductAddResult.kt
@@ -1,4 +1,4 @@
-package app.priceguard.data.dto.add
+package app.priceguard.ui.data
 
 data class ProductAddResult(
     val statusCode: Int,
